Renamed all files to lowercase ._.
