+++ demidenko.nikita/T2/main.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-#include "DataStruct.hpp"
+#include "dataStruct.hpp"
 
 int main()
 {
