+++ demidenko.nikita/T2/streamGuard.cpp
