+++ demidenko.nikita/T2/streamGuard.hpp
